chore: move BlockchainTransaction and TransactionStats files to organism folder and ensure components adapt to light and dark modes
